server side retry works with branching
