+++ .env
@@ -124,7 +124,6 @@ PUBLIC_APP_DISCLAIMER_MESSAGE="Disclaimer: AI is an area of active research with
 LLM_SUMMERIZATION=true
 
 EXPOSE_API=true
-CONVERSATION_BRANCHING=true
 # PUBLIC_APP_NAME=HuggingChat
 # PUBLIC_APP_ASSETS=huggingchat
 # PUBLIC_APP_COLOR=yellow
