+++ src/lib/components/chat/ChatWindow.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { Message } from "$lib/types/Message";
-	import { createEventDispatcher, onDestroy } from "svelte";
+	import { createEventDispatcher, onDestroy, tick } from "svelte";
 
 	import CarbonSendAltFilled from "~icons/carbon/send-alt-filled";
 	import CarbonExport from "~icons/carbon/export";
@@ -11,7 +11,6 @@
 
 	import EosIconsLoading from "~icons/eos-icons/loading";
 
-	import ChatMessages from "./ChatMessages.svelte";
 	import ChatInput from "./ChatInput.svelte";
 	import StopGeneratingBtn from "../StopGeneratingBtn.svelte";
 	import type { Model } from "$lib/types/Model";
@@ -26,6 +25,13 @@
 	import type { Assistant } from "$lib/types/Assistant";
 	import { base } from "$app/paths";
 	import ContinueBtn from "../ContinueBtn.svelte";
+	import AssistantIntroduction from "./AssistantIntroduction.svelte";
+	import ChatMessage from "./ChatMessage.svelte";
+	import ScrollToBottomBtn from "../ScrollToBottomBtn.svelte";
+	import { browser } from "$app/environment";
+	import { snapScrollToBottom } from "$lib/actions/snapScrollToBottom";
+	import SystemPromptModal from "../SystemPromptModal.svelte";
+	import ChatIntroduction from "./ChatIntroduction.svelte";
 
 	export let messages: Message[] = [];
 	export let loading = false;
@@ -96,6 +102,18 @@
 			clearTimeout(timeout);
 		}
 	});
+
+	let chatContainer: HTMLElement;
+
+	async function scrollToBottom() {
+		await tick();
+		chatContainer.scrollTop = chatContainer.scrollHeight;
+	}
+
+	// If last message is from user, scroll to bottom
+	$: if (browser && messages[messages.length - 1]?.from === "user") {
+		scrollToBottom();
+	}
 </script>
 
 <div class="relative min-h-0 min-w-0">
@@ -106,31 +124,62 @@
 			}}
 		/>
 	{/if}
-	<ChatMessages
-		{loading}
-		{pending}
-		{currentModel}
-		{models}
-		{assistant}
-		{messages}
-		readOnly={isReadOnly}
-		isAuthor={!shared}
-		{webSearchMessages}
-		{preprompt}
-		on:message={(ev) => {
-			if ($page.data.loginRequired) {
-				loginModalOpen = true;
-			} else {
-				dispatch("message", ev.detail);
-			}
-		}}
-		on:vote
-		on:continue
-		on:retry={(ev) => {
-			if (!loading) dispatch("retry", ev.detail);
-		}}
-	/>
+	<div
+		class="scrollbar-custom mr-1 h-full overflow-y-auto"
+		use:snapScrollToBottom={messages.length ? [...messages, ...webSearchMessages] : false}
+		bind:this={chatContainer}
+	>
+		<div class="mx-auto flex h-full max-w-3xl flex-col gap-6 px-5 pt-6 sm:gap-8 xl:max-w-4xl">
+			{#if $page.data?.assistant}
+				<a
+					class="mx-auto flex items-center gap-1.5 rounded-full border border-gray-100 bg-gray-50 py-1 pl-1 pr-3 text-sm text-gray-800 hover:bg-gray-100 dark:border-gray-800 dark:bg-gray-800 dark:text-gray-200 dark:hover:bg-gray-700"
+					href="{base}/settings/assistants/{$page.data.assistant._id}"
+				>
+					{#if $page.data?.assistant.avatar}
+						<img
+							src="{base}/settings/assistants/{$page.data?.assistant._id.toString()}/avatar.jpg?hash=${$page
+								.data.assistant.avatar}"
+							alt="Avatar"
+							class="size-5 rounded-full object-cover"
+						/>
+					{:else}
+						<div
+							class="flex size-6 items-center justify-center rounded-full bg-gray-300 font-bold uppercase text-gray-500"
+						>
+							{$page.data?.assistant.name[0]}
+						</div>
+					{/if}
 
+					{$page.data.assistant.name}
+				</a>
+			{:else if preprompt && preprompt != currentModel.preprompt}
+				<SystemPromptModal preprompt={preprompt ?? ""} />
+			{/if}
+
+			{#if messages.length > 0}
+				<ChatMessage
+					{loading}
+					{webSearchMessages}
+					{messages}
+					id={messages[0].id}
+					isAuthor={!shared}
+					readOnly={isReadOnly}
+					model={currentModel}
+					on:retry
+					on:vote
+					on:continue
+				/>
+			{:else if !assistant}
+				<ChatIntroduction {models} {currentModel} on:message />
+			{:else}
+				<AssistantIntroduction {assistant} on:message />
+			{/if}
+		</div>
+		<ScrollToBottomBtn
+			class="bottom-36 right-4 max-md:hidden lg:right-10"
+			scrollNode={chatContainer}
+		/>
+	</div>
 	<div
 		class="dark:via-gray-80 pointer-events-none absolute inset-x-0 bottom-0 z-0 mx-auto flex w-full max-w-3xl flex-col items-center justify-center bg-gradient-to-t from-white via-white/80 to-white/0 px-3.5 py-4 max-md:border-t max-md:bg-white sm:px-5 md:py-8 xl:max-w-4xl dark:border-gray-800 dark:from-gray-900 dark:to-gray-900/0 max-md:dark:bg-gray-900 [&>*]:pointer-events-auto"
 	>
