+++ src/routes/conversation/[id]/+page.server.ts
@@ -3,6 +3,7 @@ import { ObjectId } from "mongodb";
 import { error } from "@sveltejs/kit";
 import { authCondition } from "$lib/server/auth";
 import { UrlDependency } from "$lib/types/UrlDependency";
+import { convertLegacyConversation } from "$lib/utils/tree/convertLegacyConversation.js";
 
 export const load = async ({ params, depends, locals }) => {
 	let conversation;
@@ -45,17 +46,19 @@ export const load = async ({ params, depends, locals }) => {
 		}
 	}
 
+	const convertedConv = { ...conversation, ...convertLegacyConversation(conversation) };
+
 	return {
-		messages: conversation.messages,
-		title: conversation.title,
-		model: conversation.model,
-		preprompt: conversation.preprompt,
-		rootMessageId: conversation.rootMessageId,
-		assistant: conversation.assistantId
+		messages: convertedConv.messages,
+		title: convertedConv.title,
+		model: convertedConv.model,
+		preprompt: convertedConv.preprompt,
+		rootMessageId: convertedConv.rootMessageId,
+		assistant: convertedConv.assistantId
 			? JSON.parse(
 					JSON.stringify(
 						await collections.assistants.findOne({
-							_id: new ObjectId(conversation.assistantId),
+							_id: new ObjectId(convertedConv.assistantId),
 						})
 					)
 			  )
